backend/week4/pyproject.toml
[project]
name = "week4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-core>=1.2.18",
    "langchain-openai>=1.1.11",
    "langgraph>=1.0.10",
    "openai>=2.26.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
